+++ Imperium/src/Interface/ImperiumUI/ImperiumWindow.cs
@@ -74,7 +74,7 @@ internal abstract class ImperiumWindow : MonoBehaviour, IDragHandler, IBeginDrag
             fadeAnimation = StartCoroutine(hideAnimation());
         };
 
-        theme.OnUpdate += OnThemePrimaryUpdate;
+        theme.OnUpdate += OnThemeUpdate;
         theme.OnUpdate += value =>
         {
             ImpThemeManager.Style(
@@ -104,7 +104,7 @@ internal abstract class ImperiumWindow : MonoBehaviour, IDragHandler, IBeginDrag
         InitWindow();
 
         // Style UI with the current theme
-        OnThemePrimaryUpdate(theme.Value);
+        OnThemeUpdate(theme.Value);
 
         windowGroup.alpha = 0;
         windowGroup.interactable = false;
@@ -380,7 +380,7 @@ internal abstract class ImperiumWindow : MonoBehaviour, IDragHandler, IBeginDrag
 
     protected void CloseParent() => parent.Close();
 
-    protected virtual void OnThemePrimaryUpdate(ImpTheme themeUpdated)
+    protected virtual void OnThemeUpdate(ImpTheme themeUpdated)
     {
     }