+++ Imperium/src/Interface/ImperiumUI/Windows/ControlCenter/ControlCenterWindow.cs
@@ -22,24 +22,19 @@ internal class ControlCenterWindow : ImperiumWindow
         content = transform.Find("Content");
 
         InitPlayerSettings();
-        InitGameSettings();
 
-        RegisterWidget<LevelGeneration>(content, "Left/Generation");
+        RegisterWidget<GrabberControl>(content, "Left");
 
         if (Random.Range(0, 100) >= 99) titleBox.Find("Title").GetComponent<TMP_Text>().text = "Emporium Control Panel";
     }
 
-    protected override void OnThemePrimaryUpdate(ImpTheme themeUpdate)
+    protected override void OnThemeUpdate(ImpTheme themeUpdate)
     {
         ImpThemeManager.Style(
             themeUpdate,
             transform,
             new StyleOverride("Separator", Variant.DARKER)
         );
-    }
-
-    private void InitGameSettings()
-    {
     }
 
     private void InitPlayerSettings()